// src/libs/GOBE.d.ts
export declare interface PlayerInfo {
    playerId: string;
    status?: number;
    customPlayerStatus?: number;
    customPlayerProperties?: string;
    teamId?: string;
    isRobot?: number;
    robotName?: string;
    matchParams?: Record<string, string>;
}